test(gateway): shared invite/contact fixtures and newline-IPC client helpers (#36971)
