+++ gateway/src/__tests__/guardian-channel-create.test.ts
@@ -23,6 +23,7 @@ import {
 } from "../db/connection.js";
 import { contacts, contactChannels } from "../db/schema.js";
 import { findGuardian } from "../http/routes/guardian-channel-create.js";
+import { seedContact } from "./helpers/contact-fixtures.js";
 
 beforeAll(async () => {
   await initGatewayDb();
@@ -37,25 +38,6 @@ beforeEach(() => {
 afterAll(() => {
   resetGatewayDb();
 });
-
-function seedContact(opts: {
-  id: string;
-  role: string;
-  principalId: string | null;
-}): void {
-  const now = Date.now();
-  getGatewayDb()
-    .insert(contacts)
-    .values({
-      id: opts.id,
-      displayName: `name-${opts.id}`,
-      role: opts.role,
-      principalId: opts.principalId,
-      createdAt: now,
-      updatedAt: now,
-    })
-    .run();
-}
 
 describe("findGuardian", () => {
   test("returns the guardian seeded in the gateway DB", async () => {
